+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
-## [Unreleased]
+## [1.7.5] - 2022-08-30
 
 ### Added
 
@@ -916,7 +916,8 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 
 - Initial public release
 
-[Unreleased]: https://github.com/Snakybo/Clicked/compare/1.7.3...master
+[Unreleased]: https://github.com/Snakybo/Clicked/compare/1.7.5...master
+[1.7.5]: https://github.com/Snakybo/Clicked/releases/tag/1.7.5
 [1.7.4]: https://github.com/Snakybo/Clicked/releases/tag/1.7.4
 [1.7.3]: https://github.com/Snakybo/Clicked/releases/tag/1.7.3
 [1.7.2]: https://github.com/Snakybo/Clicked/releases/tag/1.7.2